requirements.txt
psycopg[binary]>=3.2
python-dotenv>=1.0
dbt-postgres>=1.9
